pyproject.toml
[tool.poetry]
name = "anti-uav-yolo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.6.0"
Pillow = "^9.1.1"
matplotlib = "^3.5.2"
numpy = "^1.22.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
